requirements/base.txt
aiohttp==3.8.1
alembic==1.8.1
asyncpg==0.26.0
backoff==2.1.2
elastic-apm==6.10.1
email-validator==1.2.1
fastapi==0.79.0
fastapi-pagination==0.9.3
furl==2.1.3
ipython==8.4.0
orjson==3.7.7
psycopg2-binary==2.9.3
pydantic==1.9.1
python-jose[cryptography]==3.3.0
python-json-logger==2.0.4
sentry-sdk==1.7.2
SQLAlchemy==1.4.39
typer==0.6.1
uvicorn==0.18.2
uvloop==0.16.0